+++ src/Private/Find-GitConfigPath.ps1
@@ -2,10 +2,30 @@ function Find-GitConfigPath {
   [CmdletBinding()]
   param()
 
-  # TO DO
-  # Make this more dynamic. We should be able to find this
-  # if the shell is inside a repo
-  $configPath = './.git/config'
+  $currentLocation = (Get-Location).Path
+  $expectedConfigName = '.git/config'
+  
+  # will look uproot from the current directory to find the 
+  # git config path. Max 10 folders deep
+  for($i=0;$i -le 9; $i++) {
+    # when you split path on the drive, you get a $null value back!
+    # try it: Split-Path -Parent "C:\"
+    if (-not $currentLocation) {
+      break
+    }
+
+    Write-Verbose "Looking for a git config file in $currentLocation"
+    if (Test-Path $(Join-Path $currentLocation $expectedConfigName)) {
+      $configPath = Join-Path $currentLocation $expectedConfigName
+    }
+    else {
+      $currentLocation = Split-Path -Parent $currentLocation
+    }
+  }
+
+  if (-not $configPath) {
+    throw "Could not find git config file"
+  }
 
   $configPath
 }