+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # git-open
 
+## 0.3.0
+
+-- Will now work when invoked from a folder inside of a repo by finding the git config uproot. 
+-- Adding the gitopen alias
+
 ## 0.2.0
 
 -- Added support for Azure Devops.